Avoid mangling Accept header in WinHTTP
The WinHttpOpenRequest function requires the caller to provide which
type of data to accept. wxWidgets sets this to */* to accept everything.
When a user creates a wxWebRequest and manually sets an Accept header,
then on Windows this Accept header is combined with the */* value set in
WinHttpOpenRequest. For example, when user does

request.SetHeader("Accept", "application/json");

Then the final request sends "*/*,application/json". Some web servers
only look at the first matching type and ignore the rest.

This patch checks if the caller has set an Accept header and if so,
tells the WinHttpOpenRequest to use WINHTTP_DEFAULT_ACCEPT_TYPES to
accept no type. A subsequent call to WinHttpSendRequest to set the
Accept header will fix the final request to send the proper accept type.